day05 code md

+++ day04/Src/HomeWork4.java
@@ -1,7 +1,7 @@
 class HomeWork4 {
     public static void main(String[] args) {
         /* 利用循环完成15的阶乘 */
-        int sum = 1;
+        long sum = 1;
         for (int i = 1; i<= 15; i++) {
             sum *= i;
         }
